backend/requirements.txt
django
django-auditlog
django-cors-headers
djangorestframework
drf-spectacular
gunicorn
whitenoise
djangorestframework-simplejwt